refactor(phase-48.1): split ion_writer.py from gromacs_writer (no behavior change)
- Move write_ion_gro_file + write_ion_top_file to ion_writer.py (807 lines, LARGEST)
- MoleculeIndex imported from quickice.structure_generation.types (Group 8 fix preserved)
- 2 synthetic MoleculeIndex entry sites preserved verbatim (lines 165, 176)
- custom_active gate + try/except (OSError, ValueError) cleanup + DeprecationWarning preserved
- _format_na_ion + _format_cl_ion used (unique to ion writer — ONLY writer with ion atoms)
- _write_top_defaults Format C (include_amber_header=False, compact_nbfunc_comment=False) preserved
- gromacs_writer.py re-exports both (2011 -> 1278 lines, net -733); 49 symbols preserved
- Update test_scancode_bugs_tech_debt.py: scan-all-files pattern (Path('quickice/output').glob('*.py'))
  so MoleculeIndex Group 8 regression finds all 7 sites across gromacs_writer.py (5) + ion_writer.py (2)
- 14 byte-equivalence + ion grompp + chain F1-F4 grompp + 35 MoleculeIndex regression + 15 tip4p
  scan-all-files + 127 tests/test_output ALL PASS; 1840 total tests (no loss)
